requirements.txt
streamlit
pandas
sqlglot
streamlit-authenticator
google-auth
google-cloud-storage
pyyaml
sqlalchemy
